Move `src/_components/Modal.svelte` to `$lib/components/modal.svelte`
